+++ SBMSplitMerge/NEWS.md
@@ -1,3 +1,4 @@
 # News
-## Release of version 1.0.0 - see ``README.md`` for details of the package or ``vignette("example")`` for example usage.
+## Release of version 1.1.1 - see ``README.md`` for details of the package or ``vignette("Weibull-edges")`` for example usage.
 ## Updated help pages with examples
+## Updated conjugate model sampler